+++ projects/journey-maps-client/src/lib/model/infoblock/info-block.ts
@@ -5,7 +5,7 @@ import {InfoBlockType} from './info-block-type.enum';
  * Use the matching child interface. If possible generate the info blocks with {@link InfoBlockFactoryService}
  *
  * In order to show an overlay when clicking on a marker, you need to either provide at least one InfoBlock
- * to {@link Marker#infoBlocks}, or else define a {@link JourneyMapsClientComponent#infoBoxTemplate}.
+ * to {@link Marker#infoBlocks}, or else define a {@link JourneyMapsClientComponent#markerDetailsTemplate}.
  */
 export interface InfoBlock {
   title?: string;