+++ CHANGELOG.md
@@ -6,9 +6,18 @@
 
 * DUSP-1357 Add optional `routes` input parameter to display a list of routes on the map
 
+* DUSP-1240 adds an optional `showLevelSwitch` input parameter. The level switch allows you to switch floors/levels on a
+  station map.
+
+* DUSP-1239 adds an optional `popup` input parameter. When `true`marker details will be shown in a popup instead of a
+  teaser.
+
 ### Changed
 
-* DUSP-1357 Rename journeyGeoJSON and transferGeoJSON to journey and transfer and the type from string to GeoJSON.FeatureCollection
+* **BREAKING**: DUSP-1357 Rename `journeyGeoJSON` and `transferGeoJSON` to `journey` and `transfer` and change the type
+  from `string` to `GeoJSON.FeatureCollection`
+
+* **BREAKING**: DUSP-1239 `infoBoxTemplate` input parameter has been renamed to `markerDetailsTemplate`
 
 ### Fixed